Don't force browsers to download fonts they are not going to use
I thought that browsers will make use of "type" in preload, and preload
only fonts they are going to use. Turns out this is not the case, and
browsers blindly download everything I told them to.

Preload only one font, and save almost 300kB of the page load in the process.